src/services/predictor.py
import numpy as np
from PIL import Image, ImageOps

def predict_segmentation(image_file, modality):
    """
    Dummy prediction function for brain segmentation.
    
    Args:
        image_file: UploadedFile object from Streamlit.
        modality: String, one of "CT", "CTC", or "Combined".
        
    Returns:
        PIL.Image: The original image.
        PIL.Image: The dummy segmentation mask.
    """
    # Open the image
    image = Image.open(image_file).convert("L") # Convert to grayscale
    
    # In a real scenario, we would preprocess the image, pass it to the model, and get the mask.
    # For now, we'll generate a dummy mask based on simple thresholding or random noise
    # to simulate a segmentation result.
    
    img_array = np.array(image)
    
    # Simple dummy logic: 
    # Create a mask where pixel intensity is above a certain threshold (e.g., bone/skull)
    # and add some random noise to simulate "segmentation"
    
    mask_array = np.zeros_like(img_array)
    
    # Simulate finding "regions of interest"
    threshold = 100 # Arbitrary threshold
    mask_array[img_array > threshold] = 255
    
    # Add a "tumor" or "lesion" simulation (just a circle or square in the middle)
    h, w = mask_array.shape
    center_y, center_x = h // 2, w // 2
    radius = min(h, w) // 6
    Y, X = np.ogrid[:h, :w]
    dist_from_center = np.sqrt((X - center_x)**2 + (Y-center_y)**2)
    
    # Add the "lesion" to the mask
    mask_array[dist_from_center <= radius] = 128 # Different class? Or just part of the mask
    
    # Convert back to image
    mask_image = Image.fromarray(mask_array.astype(np.uint8))
    
    return image, mask_image
