requirements.txt
torch
torchvision
torchaudio
tqdm
nibabel
numpy
matplotlib
streamlit
python-dotenv
pillow
escnn
openai
